Route housekeeping and reshuffling
