+++ src/app/multiselect/options/page.tsx
@@ -1,0 +1,26 @@
+import Link from "next/link";
+
+const LinkOption = ({
+  children,
+  href,
+}: {
+  children: React.ReactNode;
+  href: string;
+}) => (
+  <Link className="underline text-informational focusable" href={href}>
+    {children}
+  </Link>
+);
+
+export default function MultiselectOptionsPage() {
+  return (
+    <ul>
+      <li>
+        <LinkOption href="/multiselect/options/a">View Option A</LinkOption>
+      </li>
+      <li>
+        <LinkOption href="/multiselect/options/b">View Option B</LinkOption>
+      </li>
+    </ul>
+  );
+}
