+++ src/app/multiselect/success/page.tsx
@@ -1,0 +1,11 @@
+export default function MultiselectSuccessPage() {
+  return (
+    <div className="flex flex-col items-center justify-center text-titles-and-attributes">
+      <div className="flex-col max-w-lg">
+        {/* eslint-disable react/no-unescaped-entities */}
+        <h2 className="type-2xl mb-2">Success! You're done.</h2>
+        <p className="type-lg mb-8">Thanks for completing the exercise.</p>
+      </div>
+    </div>
+  );
+}
