+++ src/components/forms/types.ts
@@ -1,0 +1,5 @@
+export interface Field {
+  label: string;
+  name: string;
+  onChange?: (value: string) => void;
+}
